+++ e2e/react-start/basic/rsbuild.config.ts
@@ -1,11 +1,16 @@
 import { defineConfig } from '@rsbuild/core'
 import { pluginReact } from '@rsbuild/plugin-react'
 import { tanstackStart } from '@tanstack/react-start/plugin/rsbuild'
+import { getStartModeConfig } from './start-mode-config'
 
 const outDir = process.env.E2E_DIST_DIR ?? 'dist'
+const startModeConfig = getStartModeConfig()
 
 export default defineConfig({
-  plugins: [pluginReact({ splitChunks: false }), tanstackStart()],
+  plugins: [
+    pluginReact({ splitChunks: false }),
+    tanstackStart(startModeConfig),
+  ],
   output: {
     distPath: {
       root: outDir,
