+++ e2e/react-start/basic/vite.config.ts
@@ -2,33 +2,10 @@ import { defineConfig } from 'vite'
 import { tanstackStart } from '@tanstack/react-start/plugin/vite'
 import viteReact from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
-import { isSpaMode } from './tests/utils/isSpaMode'
-import { isPrerender } from './tests/utils/isPrerender'
-
-const spaModeConfiguration = {
-  enabled: true,
-  prerender: {
-    outputPath: 'index.html',
-  },
-}
-
-const prerenderConfiguration = {
-  enabled: true,
-  filter: (page: { path: string }) =>
-    ![
-      '/this-route-does-not-exist',
-      '/redirect',
-      '/i-do-not-exist',
-      '/not-found',
-      '/specialChars/search',
-      '/specialChars/hash',
-      '/specialChars/malformed',
-      '/users',
-    ].some((p) => page.path.includes(p)),
-  maxRedirects: 100,
-}
+import { getStartModeConfig } from './start-mode-config'
 
 const outDir = process.env.E2E_DIST_DIR ?? 'dist'
+const startModeConfig = getStartModeConfig()
 
 export default defineConfig({
   resolve: { tsconfigPaths: true },
@@ -38,12 +15,5 @@ export default defineConfig({
   server: {
     port: 3000,
   },
-  plugins: [
-    tailwindcss(),
-    tanstackStart({
-      spa: isSpaMode ? spaModeConfiguration : undefined,
-      prerender: isPrerender ? prerenderConfiguration : undefined,
-    }),
-    viteReact(),
-  ],
+  plugins: [tailwindcss(), tanstackStart(startModeConfig), viteReact()],
 })
