+++ packages/start-plugin-core/src/rsbuild/schema.ts
@@ -2,23 +2,10 @@ import { parseStartConfig as parseCoreStartConfig } from '../schema'
 import type { TanStackStartInputConfig } from '../schema'
 import type { CompileStartFrameworkOptions } from '../types'
 
-type TanStackStartRsbuildPage = Omit<
-  NonNullable<TanStackStartInputConfig['pages']>[number],
-  'prerender'
->
-
-type TanStackStartRsbuildSpa = Omit<
-  NonNullable<TanStackStartInputConfig['spa']>,
-  'prerender'
->
-
 export type TanStackStartRsbuildInputConfig = Omit<
   TanStackStartInputConfig,
-  'importProtection' | 'prerender' | 'spa' | 'pages'
-> & {
-  pages?: Array<TanStackStartRsbuildPage>
-  spa?: TanStackStartRsbuildSpa
-}
+  'importProtection'
+>
 
 export function parseStartConfig(
   opts: TanStackStartRsbuildInputConfig | undefined,
