+++ packages/start-plugin-core/tests/post-server-build.test.ts
@@ -1,5 +1,4 @@
 import { describe, expect, it, vi } from 'vitest'
-import { VITE_ENVIRONMENT_NAMES } from '../src/constants'
 
 vi.mock('@tanstack/start-server-core', () => ({
   HEADERS: {
@@ -7,32 +6,14 @@ vi.mock('@tanstack/start-server-core', () => ({
   },
 }))
 
-vi.mock('../src/vite/prerender', () => ({
-  prerender: vi.fn(async () => {}),
-}))
-
 vi.mock('../src/build-sitemap', () => ({
   buildSitemap: vi.fn(),
 }))
 
 describe('postServerBuild', () => {
-  const builder = {
-    environments: {
-      [VITE_ENVIRONMENT_NAMES.client]: {
-        config: { build: { outDir: '/client' } },
-      },
-      [VITE_ENVIRONMENT_NAMES.server]: {
-        config: { build: { outDir: '/server' } },
-      },
-    },
-    config: {
-      build: { outDir: '/root' },
-    },
-  } as any
-
   it('rejects absolute SPA maskPath URLs to avoid external prerendering', async () => {
-    // Import after mocks are set up
-    const { postServerBuild } = await import('../src/vite/post-server-build')
+    const prerender = vi.fn(async () => {})
+    const { postBuild } = await import('../src/post-build')
 
     const startConfig = {
       spa: {
@@ -47,8 +28,16 @@ describe('postServerBuild', () => {
       sitemap: { enabled: false },
     } as any
 
-    await expect(postServerBuild({ builder, startConfig })).rejects.toThrow(
-      /maskPath/i,
-    )
+    await expect(
+      postBuild({
+        startConfig,
+        adapter: {
+          getClientOutputDirectory: () => '/client',
+          prerender,
+        },
+      }),
+    ).rejects.toThrow(/maskPath/i)
+
+    expect(prerender).not.toHaveBeenCalled()
   })
 })
