+++ packages/start-plugin-core/tests/prerender-ssrf.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it, vi } from 'vitest'
-import { prerender } from '../src/vite/prerender'
-import { VITE_ENVIRONMENT_NAMES } from '../src/constants'
+import { prerender } from '../src/prerender'
 
 vi.mock('../src/utils', async () => {
   const actual = await vi.importActual<any>('../src/utils')
@@ -9,21 +8,6 @@ vi.mock('../src/utils', async () => {
     createLogger: () => ({ info: () => {}, warn: () => {}, error: () => {} }),
   }
 })
-
-vi.mock('../src/vite/prerender', async () => {
-  const actual = await vi.importActual<any>('../src/vite/prerender')
-  return {
-    ...actual,
-  }
-})
-
-// Mock vite to prevent actual server from starting
-vi.mock('vite', () => ({
-  preview: vi.fn().mockResolvedValue({
-    resolvedUrls: { local: ['http://localhost:5173/'] },
-    close: vi.fn().mockResolvedValue(undefined),
-  }),
-}))
 
 // Mock fs to prevent actual file system operations
 vi.mock('node:fs', async () => {
@@ -38,17 +22,6 @@ vi.mock('node:fs', async () => {
   }
 })
 
-const builder = {
-  environments: {
-    [VITE_ENVIRONMENT_NAMES.client]: {
-      config: { build: { outDir: '/client' } },
-    },
-    [VITE_ENVIRONMENT_NAMES.server]: {
-      config: { build: { outDir: '/server' } },
-    },
-  },
-} as any
-
 const fetchMock = vi.fn(
   async () =>
     new Response('<html></html>', {
@@ -58,6 +31,11 @@ const fetchMock = vi.fn(
 )
 
 vi.stubGlobal('fetch', fetchMock)
+
+const handler = {
+  getClientOutputDirectory: () => '/client',
+  request: fetchMock,
+}
 
 function resetFetch() {
   fetchMock.mockClear()
@@ -89,7 +67,7 @@ describe('prerender pages validation', () => {
     resetFetch()
     const startConfig = makeStartConfig('https://attacker.test/leak')
 
-    await expect(prerender({ startConfig, builder })).rejects.toThrow(
+    await expect(prerender({ startConfig, handler })).rejects.toThrow(
       /prerender page path must be relative/i,
     )
     expect(fetchMock).not.toHaveBeenCalled()
@@ -99,7 +77,7 @@ describe('prerender pages validation', () => {
     resetFetch()
     const startConfig = makeStartConfig('/about')
 
-    await expect(prerender({ startConfig, builder })).resolves.not.toThrow()
+    await expect(prerender({ startConfig, handler })).resolves.not.toThrow()
   })
 
   it('resolves when prerender filter matches no pages', async () => {
@@ -107,7 +85,7 @@ describe('prerender pages validation', () => {
     const startConfig = makeStartConfig('/about')
     startConfig.prerender.filter = () => false
 
-    await expect(prerender({ startConfig, builder })).resolves.not.toThrow()
+    await expect(prerender({ startConfig, handler })).resolves.not.toThrow()
     expect(fetchMock).not.toHaveBeenCalled()
   })
 })
